+++ Editor/UI/Analysis/Breakdowns/Shared/Comparison/Data/TreeComparisonBuilder.cs
@@ -1,7 +1,7 @@
 #if UNITY_2022_1_OR_NEWER
 using System;
 using System.Collections.Generic;
-using System.Linq;
+using Unity.MemoryProfiler.Editor.Extensions;
 using UnityEngine.UIElements;
 
 namespace Unity.MemoryProfiler.Editor.UI