+++ Editor/UI/Utilities/EditorGUICompatibilityHelper.cs
@@ -1,7 +1,4 @@
 using System;
-#if !UNITY_2021_2_OR_NEWER
-using System.Reflection;
-#endif
 using UnityEditor;
 using UnityEngine;
 using System.Collections.Generic;
@@ -13,32 +10,10 @@ namespace Unity.MemoryProfiler.Editor
         public EditorWindow window;
         public Dictionary<string, string> hyperLinkData { get; private set; }
 
-#if UNITY_2021_2_OR_NEWER
-#else
-        static PropertyInfo s_eventArghyperlinkInfos;
-#endif
-        static MemoryProfilerHyperLinkClickedEventArgs()
-        {
-#if UNITY_2021_2_OR_NEWER
-#else
-            var eventArgType = typeof(EditorGUILayout).GetNestedType("HyperLinkClickedEventArgs", BindingFlags.NonPublic);
-            s_eventArghyperlinkInfos = eventArgType.GetProperty("hyperlinkInfos", BindingFlags.Public | BindingFlags.Instance | BindingFlags.GetProperty);
-#endif
-        }
-
-#if UNITY_2021_2_OR_NEWER
         public static MemoryProfilerHyperLinkClickedEventArgs ConvertEventArguments(EditorWindow window, HyperLinkClickedEventArgs args)
         {
             return new MemoryProfilerHyperLinkClickedEventArgs() { window = window, hyperLinkData = args.hyperLinkData };
         }
-
-#else
-        public static MemoryProfilerHyperLinkClickedEventArgs ConvertEventArguments(object sender, EventArgs args)
-        {
-            return new MemoryProfilerHyperLinkClickedEventArgs() { window = EditorWindow.focusedWindow, hyperLinkData = s_eventArghyperlinkInfos.GetValue(args) as Dictionary<string, string> };
-        }
-
-#endif
     }
 
     internal static class EditorGUICompatibilityHelper
@@ -62,41 +37,23 @@ namespace Unity.MemoryProfiler.Editor
         public static event Action<MemoryProfilerHyperLinkClickedEventArgs> hyperLinkClicked = delegate { };
         const string k_hyperLinkClickedEventName = "hyperLinkClicked";
 
-#if !UNITY_2021_2_OR_NEWER
-        static EventHandler s_EventHandler;
-#endif
         static EditorGUICompatibilityHelper()
         {
-#if UNITY_2021_2_OR_NEWER
             EditorGUI.hyperLinkClicked -= OnHyperLinkClicked;
             EditorGUI.hyperLinkClicked += OnHyperLinkClicked;
-#else
-
-            var assembly = typeof(EditorGUI).Assembly;
-            var editorGUIType = typeof(EditorGUI);
-            var internalEvent = editorGUIType.GetEvent(k_hyperLinkClickedEventName, BindingFlags.NonPublic | BindingFlags.Static);
-            var eventAdder = internalEvent.GetAddMethod(true);
-            var eventRemover = internalEvent.GetRemoveMethod(true);
-            if (s_EventHandler == null)
-                s_EventHandler = OnHyperLinkClicked;
-            eventRemover.Invoke(null, new object[] { s_EventHandler });
-            eventAdder.Invoke(null, new object[] { s_EventHandler });
-#endif
         }
 
-#if UNITY_2021_2_OR_NEWER
+        [RuntimeInitializeOnLoadMethod]
+        static void RuntimeInitialize()
+        {
+            EditorGUI.hyperLinkClicked -= OnHyperLinkClicked;
+            EditorGUI.hyperLinkClicked += OnHyperLinkClicked;
+        }
+
         static void OnHyperLinkClicked(EditorWindow window, HyperLinkClickedEventArgs args)
         {
             hyperLinkClicked(MemoryProfilerHyperLinkClickedEventArgs.ConvertEventArguments(window, args));
         }
-
-#else
-        static void OnHyperLinkClicked(object sender, EventArgs args)
-        {
-            hyperLinkClicked(MemoryProfilerHyperLinkClickedEventArgs.ConvertEventArguments(sender, args));
-        }
-
-#endif
 
         public static bool DrawLinkLabel(string text, Rect rect)
         {